+++ examples/basic/src/prompt/map-chat-prompt-to-vicuna-format-example.ts
@@ -1,7 +1,7 @@
 import dotenv from "dotenv";
 import {
   LlamaCppTextGenerationModel,
-  VicunaChatPromptFormat,
+  mapChatPromptToVicunaFormat,
   streamText,
 } from "modelfusion";
 
@@ -13,7 +13,7 @@ async function main() {
     new LlamaCppTextGenerationModel({
       contextWindowSize: 2048, // Vicuna v1.5 context window size
       maxCompletionTokens: 512,
-    }).withPromptFormat(VicunaChatPromptFormat()),
+    }).withPromptFormat(mapChatPromptToVicunaFormat()),
     [
       { user: "Write a short story about a robot learning to love." },
       { ai: "Once upon a time, there was a robot who learned to love." },