+++ examples/basic/src/prompt/map-instruction-prompt-to-alpaca-format-example.ts
@@ -1,6 +1,6 @@
 import {
-  AlpacaInstructionPromptFormat,
   LlamaCppTextGenerationModel,
+  mapInstructionPromptToAlpacaFormat,
   streamText,
 } from "modelfusion";
 
@@ -10,7 +10,7 @@ async function main() {
     new LlamaCppTextGenerationModel({
       contextWindowSize: 2048, // context window size of Chronos-13B-v2-GGML
       maxCompletionTokens: 1024,
-    }).withPromptFormat(AlpacaInstructionPromptFormat()),
+    }).withPromptFormat(mapInstructionPromptToAlpacaFormat()),
     {
       instruction: "You are a celebrated poet. Write a short story about:",
       input: "a robot learning to love.",