+++ src/prompt/VicunaPromptFormat.ts
@@ -18,53 +18,55 @@ const DEFAULT_SYSTEM_PROMPT =
  * ASSISTANT:
  * ```
  */
-export const VicunaChatPromptFormat: () => PromptFormat<
+export function mapChatPromptToVicunaFormat(): PromptFormat<
   ChatPrompt,
   string
-> = () => ({
-  format: (chatPrompt) => {
-    validateChatPrompt(chatPrompt);
+> {
+  return {
+    format: (chatPrompt) => {
+      validateChatPrompt(chatPrompt);
 
-    let text = "";
+      let text = "";
 
-    for (let i = 0; i < chatPrompt.length; i++) {
-      const message = chatPrompt[i];
+      for (let i = 0; i < chatPrompt.length; i++) {
+        const message = chatPrompt[i];
 
-      // system message:
-      if (
-        i === 0 &&
-        "system" in message &&
-        typeof message.system === "string"
-      ) {
-        text += `${message.system}\n\n`;
-        continue;
+        // system message:
+        if (
+          i === 0 &&
+          "system" in message &&
+          typeof message.system === "string"
+        ) {
+          text += `${message.system}\n\n`;
+          continue;
+        }
+
+        // first message was not a system message:
+        if (i === 0) {
+          text += `${DEFAULT_SYSTEM_PROMPT}\n\n`;
+        }
+
+        // user message
+        if ("user" in message) {
+          text += `USER: ${message.user}\n`;
+          continue;
+        }
+
+        // ai message:
+        if ("ai" in message) {
+          text += `ASSISTANT:\n${message.ai}\n`;
+          continue;
+        }
+
+        // unsupported message:
+        throw new Error(`Unsupported message: ${JSON.stringify(message)}`);
       }
 
-      // first message was not a system message:
-      if (i === 0) {
-        text += `${DEFAULT_SYSTEM_PROMPT}\n\n`;
-      }
+      // AI message prefix:
+      text += `ASSISTANT: `;
 
-      // user message
-      if ("user" in message) {
-        text += `USER: ${message.user}\n`;
-        continue;
-      }
-
-      // ai message:
-      if ("ai" in message) {
-        text += `ASSISTANT:\n${message.ai}\n`;
-        continue;
-      }
-
-      // unsupported message:
-      throw new Error(`Unsupported message: ${JSON.stringify(message)}`);
-    }
-
-    // AI message prefix:
-    text += `ASSISTANT: `;
-
-    return text;
-  },
-  stopSequences: [`\nUSER:`],
-});
+      return text;
+    },
+    stopSequences: [`\nUSER:`],
+  };
+}